+++ example/test/widget_test.dart
@@ -7,6 +7,7 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:receive_sharing_intent/receive_sharing_intent.dart';
 
 import 'package:receive_sharing_intent_example/main.dart';
 
@@ -14,6 +15,12 @@ void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
     // Build our app and trigger a frame.
     await tester.pumpWidget(MyApp());
+
+    // Set mock values for ReceiveSharingIntent
+    ReceiveSharingIntent.setMockValues(
+      initialMedia: [],
+      mediaStream: Stream.empty(),
+    );
 
     // Verify that platform version is retrieved.
     expect(